requirements.txt
allure-pytest==2.8.34
allure-python-commons==2.8.34
atomicwrites==1.4.0
attrs==20.3.0
certifi==2020.12.5
chardet==4.0.0
colorama==0.4.4
comtypes==1.1.9
configparser==5.0.1
crayons==0.4.0
Faker==9.3.1
flake8==3.9.2
future==0.18.2
idna==3.1
iniconfig==1.1.1
mccabe==0.6.1
MouseInfo==0.1.3
mysql-connector-python==8.0.27
packaging==20.9
pluggy==0.13.1
protobuf==3.19.0
py==1.10.0
pycodestyle==2.7.0
pyflakes==2.3.1
PyGetWindow==0.0.9
PyMsgBox==1.0.9
PyMySQL==1.0.2
pyparsing==2.4.7
pyperclip==1.8.1
PyRect==0.1.4
PyScreeze==0.1.26
pytest==6.2.5
pytest-forked==1.3.0
pytest-xdist==2.3.0
python-dateutil==2.8.2
PyTweening==1.0.3
pywinauto==0.6.6
requests==2.15.1
selene==2.0.0a40
selenium==3.141.0
six==1.15.0
text-unidecode==1.3
toml==0.10.2
urllib3==1.26.3
webdriver-manager==3.5.0
peewee==3.14.4
adaptavist==2.0.0
pytest-env==0.6.2
pytest-xdist>=2.3.0
pytest-rerunfailures>=10.2
Pillow~=8.0.1